refactor: remove unused attributes from FoundPosition class and reorganize imports in controller.py
