+++ src/controller.py
@@ -11,12 +11,11 @@ import pyautogui
 from adbutils.errors import AdbError
 from playwright.async_api import Page
 
-from .compare import ImageComparison
+from .compare import FoundPosition, ImageComparison
 from .screenshot import Screenshot, ShiftPosition, ScreenshotManager
 from .types.config import ConfigModel
 from .utils.notify import DiscordNotify
-from .utils.get_serial import ADBDeviceManager
-from .types.output_models import FoundPosition
+from .utils.manager import ADBDeviceManager
 
 
 class RemoteController(ConfigModel):
